pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "opendocgen"
version = "0.1.0"
description = "Open-source autonomous document generation agent"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
authors = [
    {name = "OpenDocGen Contributors"}
]
keywords = ["ai", "llm", "documents", "automation", "open-source"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "crewai",
    "langchain",
    "langchain-ollama",
    "chromadb",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "pandas>=2.0.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "python-docx>=0.8.11",
    "PyPDF2>=3.0.0",
    "unstructured>=0.11.0",
    "beautifulsoup4>=4.12.0",
    "aiofiles>=23.0.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
    "mypy",
    "pre-commit",
]

[project.scripts]
opendocgen = "src.cli:main"

[project.urls]
Homepage = "https://github.com/opendocgen/opendocgen"
Documentation = "https://opendocgen.readthedocs.io"
Repository = "https://github.com/opendocgen/opendocgen"
Issues = "https://github.com/opendocgen/opendocgen/issues"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
